backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
meilisearch==0.31.0
ijson==3.2.3
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
numpy==1.26.4
# Optional — uncomment for reverse image search:
# sentence-transformers==2.5.1
# Pillow==10.2.0
# torch==2.2.1
